sdk/requirements.txt
openhands-ai>=0.14.0
ollama>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.0.0
